Move SettingsNavigation into settings folder
