+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/SettingsNavigation.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React from "react";
 import { useTranslation } from "@i18n/client";
-import { SubNavLink } from "./SubNavLink";
+import { SubNavLink } from "../../components/navigation/SubNavLink";
 import { EmailIcon, BrandIcon, GearIcon } from "@serverComponents/icons";
 
 export const SettingsNavigation = ({ id }: { id: string }) => {
